reduced deppendency on explicit calls to iotream and iomanip. using G4 methods instead.
