+++ AngraG4Simulation/vis.mac
@@ -28,6 +28,7 @@
 #/gun/position (G4ThreeVector(CLHEP::RandGauss::shoot(),CLHEP::RandGauss::shoot(),CLHEP::RandGauss::shoot()))
 /gun/particle  e-
 /gun/energy 3.5 MeV
+/gun/direction -1 0 0
 
 # beam
 /run/verbose 1
