+++ AngraG4Simulation/vis_qt.mac
@@ -27,8 +27,9 @@
 /gun/position 0 0 0 mm
 #/gun/position (G4ThreeVector(CLHEP::RandGauss::shoot(),CLHEP::RandGauss::shoot(),CLHEP::RandGauss::shoot()))
 /gun/particle  e-
-/gun/energy 3.5 MeV
+/gun/energy 1 MeV
+/gun/direction -1 0 0
 
 # beam
 /run/verbose 1
-/run/beamOn 100
+/run/beamOn 1
